spellcheck and warnings

+++ src/ui/ui_edit_doc.c
@@ -465,7 +465,7 @@ static void ui_edit_doc_copy(const ui_edit_doc_t* d,
         const int32_t c = (int32_t)(uintptr_t)(to - text);
         if (bytes > 0) {
             swear(c + bytes < b, "c: %d bytes: %d b: %d", c, bytes, b);
-            memmove(to, u, bytes);
+            memmove(to, u, (size_t)bytes);
             to += bytes;
         }
         if (pn < r.to.pn) {
@@ -488,7 +488,7 @@ static bool ui_edit_text_insert_2_or_more(ui_edit_text_t* t, int32_t pn,
     ui_edit_str_t* ps = null; // ps[np]
     bool ok = ui_edit_doc_realloc_ps_no_init(&ps, 0, np);
     if (ok) {
-        memmove(ps, t->ps, pn * sizeof(ui_edit_str_t));
+        memmove(ps, t->ps, (size_t)pn * sizeof(ui_edit_str_t));
         // `s` first line of `insert`
         ok = ui_edit_str.init(&ps[pn], s->u, s->b, true);
         // lines of `insert` between `s` and `e`
@@ -503,11 +503,12 @@ static bool ui_edit_text_insert_2_or_more(ui_edit_text_t* t, int32_t pn,
         }
         assert(t->np - pn - 1 >= 0);
         memmove(ps + pn + insert->np, t->ps + pn + 1,
-               (t->np - pn - 1) * sizeof(ui_edit_str_t));
+               (size_t)(t->np - pn - 1) * sizeof(ui_edit_str_t));
         if (ok) {
             // this two regions where moved to `ps`
             memset(t->ps, 0x00, pn * sizeof(ui_edit_str_t));
-            memset(t->ps + pn + 1, 0x00, (t->np - pn - 1) * sizeof(ui_edit_str_t));
+            memset(t->ps + pn + 1, 0x00,
+                   (size_t)(t->np - pn - 1) * sizeof(ui_edit_str_t));
             // deallocate what was copied from `insert`
             ui_edit_doc_realloc_ps_no_init(&t->ps, t->np, 0);
             t->np = np;
@@ -588,7 +589,7 @@ static bool ui_edit_text_remove_lines(ui_edit_text_t* t,
     }
     if (t->np - to - 1 > 0) {
         memmove(&t->ps[from + 1], &t->ps[to + 1],
-                (t->np - to - 1) * sizeof(ui_edit_str_t));
+                (size_t)(t->np - to - 1) * sizeof(ui_edit_str_t));
     }
     t->np -= to - from;
     if (ok) {
@@ -679,7 +680,7 @@ static void ui_edit_text_copy(const ui_edit_text_t* t,
         const int32_t c = (int32_t)(uintptr_t)(to - text);
         swear(c + bytes < b, "d: %d bytes:%d b: %d", c, bytes, b);
         if (bytes > 0) {
-            memmove(to, u, bytes);
+            memmove(to, u, (size_t)bytes);
             to += bytes;
         }
         if (pn < r.to.pn) {
@@ -1114,7 +1115,7 @@ static void ui_edit_str_free(ui_edit_str_t* s) {
 static bool ui_edit_str_init_g2b(ui_edit_str_t* s) {
     const int64_t _4_bytes = (int64_t)sizeof(int32_t);
     // start with number of glyphs == number of bytes (ASCII text):
-    bool ok = ut_heap.alloc(&s->g2b, (s->b + 1) * _4_bytes) == 0;
+    bool ok = ut_heap.alloc(&s->g2b, (size_t)(s->b + 1) * _4_bytes) == 0;
     int32_t i = 0; // index in u[] string
     int32_t k = 1; // glyph number
     // g2b[k] start postion in uint8_t offset from utf8 text of glyph[k]
@@ -1157,7 +1158,7 @@ static bool ui_edit_str_init(ui_edit_str_t* s, const char* u, int32_t b,
     } else {
         if (heap) {
             ok = ut_heap.alloc((void**)&s->u, b) == 0;
-            if (ok) { s->c = b; memmove(s->u, u, b); }
+            if (ok) { s->c = b; memmove(s->u, u, (size_t)b); }
         } else {
             s->u = (char*)u;
         }
@@ -1198,7 +1199,7 @@ static bool ui_edit_str_move_g2b_to_heap(ui_edit_str_t* s) {
         }
         const int32_t bytes = (s->g + 1) * (int32_t)sizeof(int32_t);
         ok = ut_heap.alloc(&s->g2b, bytes) == 0;
-        if (ok) { memmove(s->g2b, ui_edit_str_g2b_ascii, bytes); }
+        if (ok) { memmove(s->g2b, ui_edit_str_g2b_ascii, (size_t)bytes); }
     }
     return ok;
 }
@@ -1209,7 +1210,7 @@ static bool ui_edit_str_move_to_heap(ui_edit_str_t* s, int32_t c) {
     if (s->c == 0) { // s->u points outside of the heap
         const char* o = s->u;
         ok = ut_heap.alloc((void**)&s->u, c) == 0;
-        if (ok) { memmove(s->u, o, s->b); }
+        if (ok) { memmove(s->u, o, (size_t)s->b); }
     } else if (s->c < c) {
         ok = ut_heap.realloc((void**)&s->u, c) == 0;
     }
@@ -1269,9 +1270,10 @@ static bool ui_edit_str_remove(ui_edit_str_t* s, int32_t f, int32_t t) {
         if (ok) {
             const int32_t bytes_to_shift = s->b - s->g2b[t];
             assert(0 <= bytes_to_shift && bytes_to_shift <= s->b);
-            memmove(s->u + s->g2b[f], s->u + s->g2b[t], bytes_to_shift);
+            memmove(s->u + s->g2b[f], s->u + s->g2b[t], (size_t)bytes_to_shift);
             if (s->g2b != ui_edit_str_g2b_ascii) {
-                memmove(s->g2b + f, s->g2b + t, (s->g - t + 1) * sizeof(int32_t));
+                memmove(s->g2b + f, s->g2b + t,
+                        (size_t)(s->g - t + 1) * sizeof(int32_t));
                 for (int32_t i = f; i <= s->g; i++) {
                     s->g2b[i] -= bytes_to_remove;
                 }
@@ -1325,16 +1327,16 @@ static bool ui_edit_str_replace(ui_edit_str_t* s,
                 if (bytes_to_insert <= bytes_to_remove) {
                     memmove(s->u + s->g2b[f] + bytes_to_insert,
                            s->u + s->g2b[f] + bytes_to_remove,
-                           s->b - s->g2b[f] - bytes_to_remove);
+                           (size_t)(s->b - s->g2b[f] - bytes_to_remove));
                     if (all_ascii) {
                         assert(s->g2b == ui_edit_str_g2b_ascii);
                     } else {
                         assert(s->g2b != ui_edit_str_g2b_ascii);
                         memmove(s->g2b + f + glyphs_to_insert,
                                s->g2b + f + glyphs_to_remove,
-                               (s->g - t + 1) * _4_bytes);
+                               (size_t)(s->g - t + 1) * _4_bytes);
                     }
-                    memmove(s->u + s->g2b[f], ins.u, ins.b);
+                    memmove(s->u + s->g2b[f], ins.u, (size_t)ins.b);
                 } else {
                     if (all_ascii) {
                         assert(s->g2b == ui_edit_str_g2b_ascii);
@@ -1343,22 +1345,23 @@ static bool ui_edit_str_replace(ui_edit_str_t* s,
                         const int32_t g = s->g + glyphs_to_insert -
                                                  glyphs_to_remove;
                         assert(g > s->g);
-                        ok = ut_heap.realloc(&s->g2b, (g + 1) * _4_bytes) == 0;
+                        ok = ut_heap.realloc(&s->g2b,
+                                             (size_t)(g + 1) * _4_bytes) == 0;
                     }
                     // need to shift bytes staring with s.g2b[t] toward the end
                     if (ok) {
                         memmove(s->u + s->g2b[f] + bytes_to_insert,
                                 s->u + s->g2b[f] + bytes_to_remove,
-                                s->b - s->g2b[f] - bytes_to_remove);
+                                (size_t)(s->b - s->g2b[f] - bytes_to_remove));
                         if (all_ascii) {
                             assert(s->g2b == ui_edit_str_g2b_ascii);
                         } else {
                             assert(s->g2b != ui_edit_str_g2b_ascii);
                             memmove(s->g2b + f + glyphs_to_insert,
                                     s->g2b + f + glyphs_to_remove,
-                                    (s->g - t + 1) * _4_bytes);
+                                    (size_t)(s->g - t + 1) * _4_bytes);
                         }
-                        memmove(s->u + s->g2b[f], ins.u, ins.b);
+                        memmove(s->u + s->g2b[f], ins.u, (size_t)ins.b);
                     }
                 }
                 if (ok) {
@@ -1628,7 +1631,7 @@ static void ui_edit_doc_test_big_text(void) {
     enum { MB10 = 10 * 1000 * 1000 };
     char* text = null;
     ut_heap.alloc(&text, MB10);
-    memset(text, 'a', MB10 - 1);
+    memset(text, 'a', (size_t)MB10 - 1);
     char* p = text;
     uint32_t seed = 0x1;
     for (;;) {
